Move all clap args to main module
